requirement.txt
asgiref==3.8.1
Babel==2.15.0
crispy-bootstrap5==2024.2
Django==4.2.14
django-appointment==3.6.0
django-crispy-forms==2.3
django-phonenumber-field==7.3.0
django-picklefield==3.2
django-q2==1.6.2
django-scheduler==0.10.1
djangorestframework==3.15.2
icalendar==5.0.13
importlib_metadata==8.2.0
phonenumbers==8.13.42
pillow==10.4.0
pkg_resources==0.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
zipp==3.19.2
